🔧 Update configure.sh with safety checks and bin/ linking

+++ configure.sh
@@ -2,10 +2,17 @@
 
 script_dir="$(dirname "$(realpath "${BASH_SOURCE[0]}")")"
 
-for app in mimeapps.list system-images fontconfig foot ghostty waybar sway swaylock rofi dunst cmus opencode; do
+mkdir -p $HOME/.config 
+mkdir -p $HOME/.local/bin
+
+for app in mimeapps.list system-images foot ghostty waybar sway swaylock rofi dunst cmus opencode cava; do
   src="$script_dir/$app"
   tgt="$HOME/.config/$app"
-  rm -rf "$tgt"
+
+  if [[ -e "$tgt" ]]; then
+      echo "⚠ WARNING: $tgt already exists — not replacing"
+      continue
+  fi
   ln -sf "$src" "$tgt"
   echo "Linked $src   $tgt"
 done
@@ -13,10 +20,32 @@ done
 rm -rf "$script_dir/rofi/nerd-fonts"
 source "$script_dir/rofi/download-icons.sh"
 
+
+# Add scripts to .local/bin
+for app in bin/*; do
+    [[ -f "$app" ]] || continue
+
+    src="$script_dir/$app"
+    tgt="$HOME/.local/$app"
+
+    if [[ -e "$tgt" ]]; then
+        echo "⚠ WARNING: $tgt already exists — not replacing"
+        continue
+    fi
+
+    ln -s "$src" "$tgt"
+    echo "Linked $src → $tgt"
+done
+
 for conf in .bash_profile; do
   src="$script_dir/$conf"
   tgt="$HOME/$conf"
-  rm -rf "$tgt"
+
+  if [[ -e "$tgt" ]]; then
+      echo "⚠ WARNING: $tgt already exists — not replacing"
+      continue
+  fi
+
   ln -sf "$src" "$tgt"
   echo "Linked $src   $tgt"
 done
